refactor: Separate EquipmentManagementSystem class from Character
