+++ Assets/Script/Equipment.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: fdf15ddf3d6e7e6409f5f90092e29f6d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
