+++ Assets/Script/Equipment/EquipmentManagementSystem.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 924f6dc1678d2ac4ab755258eb4131c1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
