+++ Assets/Script/Equipment/Weapon.meta
