+++ Assets/Script/Equipment/Weapon/Weapon.cs.meta
