+++ Assets/Script/Player/Character.cs
@@ -12,7 +12,7 @@ using System.Linq;
 using Unity.VisualScripting.Antlr3.Runtime;
 using static UnityEditor.IMGUI.Controls.PrimitiveBoundsHandle;
 
-public class Character : MonoBehaviour
+public class Character : EquipmentManagementSystem
 {
     //캐릭터의 스탯지정
     //예시를 위해 값은 무작위로 넣음
@@ -28,36 +28,24 @@ public class Character : MonoBehaviour
     private int mLevel;
     private int mExp;
     private int mMaxExp;
-    private int mdExp;
 
     public float[] CharacterStats;
-    public List<Weapon> Weapons;
-    public List<Accessory> Accessories;
-    public List<int> MasteredWeapons;
-    public List<int> MasteredAccessories;
     public RandomPickUpSystem RandomPickUpSystem;
 
-    public int[] TransWeaponIndex; // 해당 index의 weapon이 현재 보유중인 Weapons의 몇 번째 index에 있는지 반환하는 배열, 없다면 -1 반환
-    public int[] TransAccessoryIndex; // 위와 같으나 Accessory에 해당
     void Start()
     {
         mLevel = 1;
         mExp = 0;
         mMaxExp = 100;
-        mdExp = 10;
 
         // TODO: user가 메인 화면에서 강화해놓은 스탯들을 기본값으로 받아오기
         CharacterStats = new float[21] { 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 70, 12, 13, 14, 15, 16, 17, 18, 19, 20 };
-
         Weapons = new List<Weapon>();
         Accessories = new List<Accessory>();
-
         TransWeaponIndex = Enumerable.Repeat<int>(-1, 13).ToArray<int>();
         TransAccessoryIndex = Enumerable.Repeat<int>(-1, 21).ToArray<int>();
-
         MasteredWeapons = new List<int>();
         MasteredAccessories = new List<int>();
-
         RandomPickUpSystem = new RandomPickUpSystem();
         UpdateLuck(CharacterStats[(int)Enums.Stat.Luck]);
 
@@ -82,7 +70,7 @@ public class Character : MonoBehaviour
         while (mExp >= mMaxExp)
         {
             mExp -= mMaxExp;
-            mMaxExp += mdExp;
+            mMaxExp += Constants.DeltaExp;
             LevelUp();
         }
     }
@@ -98,88 +86,9 @@ public class Character : MonoBehaviour
     }
     public void UpdateLuck(float luck)
     {
-        RandomPickUpSystem.UpdateWeaponPickUpList(luck, this);
-        RandomPickUpSystem.UpdateAccessoryPickUpList(luck, this);
+        RandomPickUpSystem.UpdateWeaponPickUpList(this);
+        RandomPickUpSystem.UpdateAccessoryPickUpList(this);
     }
-    
-    public void ApplyItem(Tuple<int, int, int> pickUp)
-    {
-        switch ((Enums.PickUpType)pickUp.Item1)
-        {
-            case Enums.PickUpType.Weapon:
-                applyWeapon(pickUp.Item2, pickUp.Item3);
-                break;
-            case Enums.PickUpType.Accessory:
-                applyAccessory(pickUp.Item2, pickUp.Item3);
-                break;
-            default:
-                applyEtc(pickUp.Item2);
-                break;
-        }
-    }
-    private void applyWeapon(int weaponIndex, int hasWeapon)
-    {
-        if (hasWeapon == 0)
-            GetWeapon(weaponIndex);
-        else
-            UpgradeWeapon(weaponIndex);
-    }
-    private void applyAccessory(int accessoryIndex, int hasAccessory)
-    {
-        if (hasAccessory == 0)
-            GetAccessory(accessoryIndex);
-        else
-            UpgradeAccessory(accessoryIndex);
-    }
-    private void applyEtc(int etcIndex)
-    {
-        switch ((Enums.Etc)etcIndex)
-        {
-            case Enums.Etc.Food:
-                // TODO: 체력 회복 함수와 연결
-                break;
-            case Enums.Etc.Money:
-                // TODO: 재화 획득 함수와 연결
-                break;
-            default:
-                break;
-        }
-    }
-    //ToDo: SkillFiringSystem이랑 연계 할 함수
-    public void GetWeapon(int weaponIndex)
-    {
-        TransWeaponIndex[weaponIndex] = Weapons.Count;
-        Weapon newWeapon = this.gameObject.AddComponent<Weapon>();
-        newWeapon.WeaponSetting(weaponIndex);
-        Weapons.Add(newWeapon);
-        RandomPickUpSystem.UpdateWeaponPickUpList(CharacterStats[(int)Enums.Stat.Luck], this);
-    }
-    public void UpgradeWeapon(int weaponIndex)
-    {
-        Weapons[TransWeaponIndex[weaponIndex]].Upgrade();
-        if (Weapons[TransWeaponIndex[weaponIndex]].IsMaster())
-        {
-            MasteredWeapons.Add(weaponIndex);
-            RandomPickUpSystem.UpdateWeaponPickUpList(CharacterStats[(int)Enums.Stat.Luck], this);
-        }
-    }
-    public void GetAccessory(int accessoryIndex)
-    {
-        TransAccessoryIndex[accessoryIndex] = Accessories.Count;
-        Accessories.Add(new Accessory(accessoryIndex));
-        UpgradeAccessory(accessoryIndex);
-        RandomPickUpSystem.UpdateAccessoryPickUpList(CharacterStats[(int)Enums.Stat.Luck], this);
-    }
-    public void UpgradeAccessory(int accessoryIndex)
-    {
-        Accessories[TransAccessoryIndex[accessoryIndex]].Upgrade(this);
-        if (Accessories[TransAccessoryIndex[accessoryIndex]].IsMaster())
-        {
-            MasteredAccessories.Add(accessoryIndex);
-            RandomPickUpSystem.UpdateAccessoryPickUpList(CharacterStats[(int)Enums.Stat.Luck], this);
-        }
-    }
-
     //Get,Set함수 자동 구현
     public int Damage
     {
