+++ Assets/Script/System/RandomPickUpSystem.cs
@@ -17,13 +17,9 @@ public class RandomPickUpSystem
         mWeaponRarity = new int[13] { 100, 100, 100, 100, 80, 80, 80, 70, 100, 50, 50, 80, 80 };
         mAccessoryRarity = new int[21] { 100, 100, 100, 90, 90, 90, 80, 80, 80, 70, 70, 70, 60, 60, 60, 50, 50, 50, 40, 40, 40 }; // юс╫ц
     }
-    public RandomPickUpSystem()
+    public void UpdateWeaponPickUpList(Character character)
     {
-
-    }
-    public void UpdateWeaponPickUpList(float luck, Character character)
-    {
-        character.CharacterStats[(int)Enums.Stat.Luck] = luck;
+        var luck = character.CharacterStats[(int)Enums.Stat.Luck];
         mWeaponPicker = new WeightedRandomPicker<int>();
         if (character.Weapons.Count < Constants.MaxWeaponCount)
         {
@@ -45,8 +41,9 @@ public class RandomPickUpSystem
             }
         }
     }
-    public void UpdateAccessoryPickUpList(float luck, Character character)
+    public void UpdateAccessoryPickUpList(Character character)
     {
+        var luck = character.CharacterStats[(int)Enums.Stat.Luck];
         mAccessoryPicker = new WeightedRandomPicker<int>();
         if (character.Accessories.Count < Constants.MaxAccessoryCount)
         {
